Feature: Item x Item Comparison UI (#40)
- [x] Add custom view VirtusizeProductImageView to easily display the product image on InPage
- [x] Fix displaying the wrong recommended size info
- [x] Improve error handling
- [x] Display product thumbnails with crossfade animation when the width of InPage is below 411dp
- [x] Handle different scenarios when an event, such as "user-selected-product", "user-added-product", "user-logged-in", "user-logged-out" and so on, is sent from Aoyama to SDK
- [x] Handle events outside of the for loop of messageHandlers
- [x] Fix the vertical margins of user and store product image views being too big on a small screen
- [x] Overlap product images with a smaller margin
- [x] Fix the wrong tint color of the store product image view

+++ sampleAppKotlin/src/main/java/com/virtusize/android/MainActivity.kt
@@ -3,7 +3,6 @@ package com.virtusize.android
 import android.os.Bundle
 import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
 import com.virtusize.libsource.data.local.*
 import com.virtusize.libsource.data.local.VirtusizeOrder
 import com.virtusize.libsource.ui.VirtusizeView
@@ -45,7 +44,7 @@ class MainActivity : AppCompatActivity() {
         exampleVirtusizeInPageStandard.virtusizeViewStyle = VirtusizeViewStyle.TEAL
         // If you like, you can set up the horizontal margins between the edges of the app screen and the InPage Standard view
         // Note: Use the helper extension function `dpInPx` if you like
-        exampleVirtusizeInPageStandard.horizontalMargin = 16.dpInPx.toFloat()
+        exampleVirtusizeInPageStandard.horizontalMargin = 16.dpInPx
         /*
          * If you like, you can set up the background color of the check size button in InPage Standard,
          * as long as it passes WebAIM contrast test.
